IC-64 - refactoring
