--- src/primary/ipc/pipe/mod.rs
@@ -1,1 +1,0 @@
-pub mod types;
