+++ src/primary/shared/mod.rs
@@ -1,3 +1,2 @@
-pub mod pipe;
 pub mod session;
 pub mod storage;
